refactor: websocket logic
