+++ server/api/studio/pkg/ws/utils/hub.go
@@ -1,4 +1,4 @@
-package ws
+package utils
 
 // Hub maintains the set of active clients and broadcasts messages to the
 // clients.
