+++ server/api/studio/pkg/ws/utils/middleware.go
@@ -1,0 +1,5 @@
+package utils
+
+type TNext func(msg *MessageReceive, client *Client) *MessagePost
+
+type TMiddleware func(next TNext) TNext
